+++ pools/usecase/pools_usecase_test.go
@@ -18,6 +18,7 @@ import (
 	"github.com/osmosis-labs/osmosis/v25/x/gamm/pool-models/stableswap"
 	gammtypes "github.com/osmosis-labs/osmosis/v25/x/gamm/types"
 	poolmanagertypes "github.com/osmosis-labs/osmosis/v25/x/poolmanager/types"
+	cosmwasmdomain "github.com/osmosis-labs/sqs/domain/cosmwasm"
 	passthroughdomain "github.com/osmosis-labs/sqs/domain/passthrough"
 
 	"github.com/osmosis-labs/sqs/domain"
@@ -102,7 +103,7 @@ func (s *PoolsUsecaseTestSuite) TestGetRoutesFromCandidates() {
 	brokenChainPool := *defaultPool
 	brokenChainPool.PoolType = poolmanagertypes.CosmWasm
 
-	cosmWasmPoolsParams := pools.CosmWasmPoolsParams{
+	cosmWasmPoolsParams := cosmwasmdomain.CosmWasmPoolsParams{
 		ScalingFactorGetterCb: domain.UnsetScalingFactorGetterCb,
 	}
 	_, err = pools.NewRoutablePool(&brokenChainPool, denomTwo, defaultTakerFee, cosmWasmPoolsParams)
@@ -942,7 +943,7 @@ func (s *PoolsUsecaseTestSuite) TestRetainPoolIfMatchesOptions() {
 }
 
 func (s *PoolsUsecaseTestSuite) newRoutablePool(pool sqsdomain.PoolI, tokenOutDenom string, takerFee osmomath.Dec) domain.RoutablePool {
-	cosmWasmPoolsParams := pools.CosmWasmPoolsParams{
+	cosmWasmPoolsParams := cosmwasmdomain.CosmWasmPoolsParams{
 		ScalingFactorGetterCb: domain.UnsetScalingFactorGetterCb,
 	}
 	routablePool, err := pools.NewRoutablePool(pool, tokenOutDenom, takerFee, cosmWasmPoolsParams)