+++ CHANGELOG.md
@@ -38,6 +38,9 @@ Ref: https://keepachangelog.com/en/1.0.0/
 ## Unreleased
 
 - Orderbook plugin
+- Change tick iteration exit condition in CL in orderbook from zero to smallest Dec (18 decimals).
+- Add rounding direction to the order book
+- Fix bug in skipping Numia API response errors 
 
 ## v25.9.0